Add files via upload
Consists of 2 problems.
Problem 1: User enters a "base" and an "exponent". The program calculates the number and sums its digits. Repeats summing the digits until the summation is a digit.
Problem 2: User enters a sequence of sequential numbers starting from 1. ( for example, "1,2,3,4,5,6,7") The program converts the sequence into a "lucky number" sequence.