Add files via upload
Quiz 5: We are given 2 input files like operands.txt and comparison_data.txt. For each line in operands.txt, the program should find all integers in between the third and the fourth integer that are divisible by the first integer but not by the second one.
The program should then compare its results with the results in the comparison_data.txt and produce an AssertionError if they don't match. Also, the program should handle bad inputs instead of crashing.